+++ changelog.md
@@ -11,3 +11,8 @@ All notable changes to `LaravelTwilio` will be documented in this file.
 
 ### Added
 - Added `SMSChannel Class`
+
+## Version 5.0.0
+### Added
+- Laravel Notification has been refactored. Detail at [README.md](/readme.md#laravelnotification)
+- Sandbox Mode. Detail at [README.md](/readme.md#sandbox-mode)